+++ model/employee.go
@@ -15,16 +15,16 @@ type GetOrdersResult struct {
 	OrderName       string    `json:"order_name"`
 	CustomerName    string    `json:"customer_name"`
 	CustomerCompany string    `json:"customer_company"`
-	DeliveredAmount *uint     `json:"delivered_amount"`
-	TotalAmount     uint      `json:"total_amount"`
+	DeliveredAmount *float32  `json:"delivered_amount"`
+	TotalAmount     float32   `json:"total_amount"`
 }
 
 type CreateOrderRequest struct {
-	OrderName       string `json:"order_name" validate:"required,notblank,alphanum,min=3,max=60"`
-	CustomerName    string `json:"customer_name" validate:"required,notblank,min=3,max=60"`
-	CustomerCompany string `json:"customer_company" validate:"required,notblank,min=3,max=60"`
-	TotalAmount     uint   `json:"total_amount" validate:"required,numeric"`
-	DeliveredAmount uint   `json:"delivered_amount" validate:"omitempty,numeric"`
+	OrderName       string  `json:"order_name" validate:"required,notblank,alphanum,min=3,max=60"`
+	CustomerName    string  `json:"customer_name" validate:"required,notblank,min=3,max=60"`
+	CustomerCompany string  `json:"customer_company" validate:"required,notblank,min=3,max=60"`
+	TotalAmount     float32 `json:"total_amount" validate:"required,numeric"`
+	DeliveredAmount float32 `json:"delivered_amount" validate:"omitempty,numeric"`
 }
 
 type CreateOrderResult struct {
@@ -33,8 +33,8 @@ type CreateOrderResult struct {
 	OrderName       string    `json:"order_name"`
 	CustomerName    string    `json:"customer_name"`
 	CustomerCompany string    `json:"customer_company"`
-	DeliveredAmount *uint     `json:"delivered_amount"`
-	TotalAmount     uint      `json:"total_amount"`
+	DeliveredAmount *float32  `json:"delivered_amount"`
+	TotalAmount     float32   `json:"total_amount"`
 }
 
 type GetOrderByIDRequest struct {
@@ -51,18 +51,18 @@ type GetOrderByIDResult struct {
 	OrderName       string    `json:"order_name"`
 	CustomerName    string    `json:"customer_name"`
 	CustomerCompany string    `json:"customer_company"`
-	DeliveredAmount *uint     `json:"delivered_amount"`
-	TotalAmount     uint      `json:"total_amount"`
+	DeliveredAmount *float32  `json:"delivered_amount"`
+	TotalAmount     float32   `json:"total_amount"`
 }
 
 type EditOrderRequest struct {
 	OrderID int `param:"id" validate:"required"` // Path variable
 
-	OrderName       string `json:"order_name" validate:"required,notblank,alphanum,min=3,max=60"`
-	CustomerName    string `json:"customer_name" validate:"required,notblank,min=3,max=60"`
-	CustomerCompany string `json:"customer_company" validate:"required,notblank,min=3,max=60"`
-	TotalAmount     uint   `json:"total_amount" validate:"required,numeric"`
-	DeliveredAmount uint   `json:"delivered_amount" validate:"omitempty,numeric"`
+	OrderName       string  `json:"order_name" validate:"required,notblank,alphanum,min=3,max=60"`
+	CustomerName    string  `json:"customer_name" validate:"required,notblank,min=3,max=60"`
+	CustomerCompany string  `json:"customer_company" validate:"required,notblank,min=3,max=60"`
+	TotalAmount     float32 `json:"total_amount" validate:"required,numeric"`
+	DeliveredAmount uint    `json:"delivered_amount" validate:"omitempty,numeric"`
 }
 
 type EditOrderResult struct {
@@ -71,6 +71,6 @@ type EditOrderResult struct {
 	OrderName       string    `json:"order_name"`
 	CustomerName    string    `json:"customer_name"`
 	CustomerCompany string    `json:"customer_company"`
-	DeliveredAmount *uint     `json:"delivered_amount"`
-	TotalAmount     uint      `json:"total_amount"`
+	DeliveredAmount *float32  `json:"delivered_amount"`
+	TotalAmount     float32   `json:"total_amount"`
 }
